requirements.txt
numpy
pandas
opencv
Pillow
albumentations
tensorflow
matplotlib
